// 前端的PrintInterface【可参考】/types.ts
export type ScaleOption = "noscale" | "shrink" | "fit"
export type MethodOption = "duplex" | "duplexshort" | "simplex"
export type PaperOption = "A2" | "A3" | "A4" | "A5" | "A6" | "letter" | "legal" | "tabloid"
export type OrientationOption = "portrait" | "landscape"
// ...
export type PrintType = "text" | "barCode" | "qrCode" | "image" | "table";
export type PrintPosition = "left" | "center" | "right";
export type PrintTableField = {
	type: "text" | "image";
	value?: string;
	path?: string;
	style?: PrintStyle;
	width?: string;
	height?: string;
}
export type PrintData = {
    type: PrintType;
	value?: string;
	style?: PrintStyle;
	width?: number;
	height?: number;
	fontsize?: number;
	displayValue?: boolean;
	position?: PrintPosition;
	path?: string;
	url?: string;
	tableHeader?: PrintTableField[] | string[];
	tableBody?: PrintTableField[][] | string[][];
	tableFooter?: PrintTableField[] | string[];
	tableHeaderStyle?: PrintStyle;
	tableBodyStyle?: PrintStyle;
	tableFooterStyle?: PrintStyle;
}

export type PrintMargin = {
    marginType?: 'default' | 'none' | 'printableArea' | 'custom';
    top?: number;
    bottom?: number;
    right?: number;
    left?: number;
}

export interface SizeOptions {
    height: number;
    width: number;
}

export type Printer = {
    id: string;
    name: string;
    driver_name: string;
    job_count: number;
    print_processor: string;
    port_name: string;
    share_name: string;
    computer_name: string;
    printer_status: number;  // https://learn.microsoft.com/en-us/windows/win32/cimwin32prov/win32-printer
    shared: boolean;
    type: number; // 0: local; 1: connection
    priority: number
}

export type ColorType = "color" | "monochrome"

export type RangeOptions = {
    from: number,
    to: number
}
export type PrintSettings = {
    paper?: PaperOption;
    method?: MethodOption;
    scale?: ScaleOption;
    color_type?: ColorType
    orientation?: OrientationOption;
    repeat?: number;
    range?: RangeOptions|string
}
export type PrintOptions = {
    id?: string;
    name?: string;
    preview?: boolean;
    page_size: SizeOptions;
    print_setting?: PrintSettings;
    remove_temp?: boolean;
}

export type PrintFileOptions = {
    id?: string;
    name?: string;
    path?: string;
    file?: Buffer,
    print_setting?: PrintSettings;
    remove_temp?: boolean;
}

export type JobsStatus = {
    code: number | string;
    name: string;
	description: string;
	error?: any,
	errorDetails?: any
}
export type Jobs = {
    id: string,
    size: number,
    printer_name: string,
    job_status: JobsStatus | null,
    computer_name?: string,
    data_type?: string,
    document_name?: string,
    job_id?: string,
    job_time?: number,
    pages_printed?: number,
    position?: number,
    priority?: number,
    submitted_time?: number | null,
    total_pages?: number,
    username?: string,
	taskID?: string,
    create_time?: number | null,
	printer_id?: string | null,
	jobs?: any[],
	request_id: string,
	task_id: string,
	documents?: {
		contents: any[]
		documentID: string
	}[],
	driver_name?: string,
	print_processor?: string,

}

export type ResponseResult = {
    message: string|undefined;
    success: boolean;
}